+++ pkg/scheduler/factory/factory.go
@@ -20,8 +20,6 @@ package factory
 
 import (
 	"fmt"
-	"os"
-	"os/signal"
 	"reflect"
 	"time"
 
@@ -385,28 +383,18 @@ func NewConfigFactory(args *ConfigFactoryArgs) Configurator {
 		},
 	)
 
-	// Setup cache comparer
+	// Setup cache debugger
 	debugger := cachedebugger.New(
 		args.NodeInformer.Lister(),
 		args.PodInformer.Lister(),
 		c.schedulerCache,
 		c.podQueue,
 	)
-
-	ch := make(chan os.Signal, 1)
-	signal.Notify(ch, compareSignal)
+	debugger.ListenForSignal(c.StopEverything)
 
 	go func() {
-		for {
-			select {
-			case <-c.StopEverything:
-				c.podQueue.Close()
-				return
-			case <-ch:
-				debugger.Comparer.Compare()
-				debugger.Dumper.DumpAll()
-			}
-		}
+		<-c.StopEverything
+		c.podQueue.Close()
 	}()
 
 	return c
