+++ pkg/scheduler/internal/cache/debugger/signal.go
@@ -1,3 +1,5 @@
+// +build !windows
+
 /*
 Copyright 2018 The Kubernetes Authors.
 
@@ -14,10 +16,10 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package factory
+package debugger
 
-import "os"
+import "syscall"
 
-// compareSignal is the signal to trigger cache compare. For windows,
-// it's SIGINT.
-var compareSignal = os.Interrupt
+// compareSignal is the signal to trigger cache compare. For non-windows
+// environment it's SIGUSR2.
+var compareSignal = syscall.SIGUSR2
